Create index.php
main page of Invisible Creations.  PHP is used to display a sample of eight photographs.  Basic calculations are performed and the results are stored in PHP variables, which are then merged with HTML output for a clean and dynamic layout.

CSS animations are used in conjunction with Javascript to animate the eight photographs, and to allow a user to select any photo they'd like in order to make it the main image in the center.  Animation scales between 768px width and 1320px width.  Below 768px, the animation is changed completely to a different layout and animation type.