Create contactValidation.php
This file pulls post data from the form on the contact page and first validates that the data is populated into each required field.  If so, it then validates that the data from each field is properly formatted based on the data type (email is an email, names don't contain numbers or special characters, website is a valid website, etc.).

If any errors occur, an error message is generated and will show on the contact page form, describing how to fix the error.  If no errors are generated, two emails are then created, one to notify myself of a communication, the other to confirm to the user that their message has been sent for review.  The emails are sent using the PHP mail() function.