Create main.js
This is the main Javascript file which is included in each webpage.  Four Event Listeners are added (so long as the element ID or class exists on the current webpage).

The first event listener is used to expand the mobile navigation, only if the screen is below 768px.

The next event listener calls a function for the home page.  If one of the animated images is clicked, it will become the central image on that page and will cease part of the page's animation (the central image will no longer change every 4 seconds).

The other two event listeners call functions for the gallery page.  If an image on the gallery page is clicked, that image will enlarge, using CSS animations, to fill the page.  If a close button on the image is then clicked, the image will shrink back down to the normal gallery page layout.