Create gallery.php
The gallery.php file is the gallery page.  Using PHP, all JPG files are imported from the folder, galleryImages.  The images are resized and using PHP the proper margins and classes are calculated to ensure a clean and dynamic layout.  With the use of PHP, this page does not require any code updates when new photographs are uploaded to the galleryImages folder.  They will be automatically added and formatted to the page.

CSS animations are used on this page for enlargement/closure of a selected photograph, and also when the cursor hovers over an image before selecting it.