Create contact.php
The contact page includes basic contact information and an HTML form to send a message.  If a user selects to use the form, it collects and validates first and last name, email address, website url (which is optional), and the message/question.  All information is validated through the contactValidation.php file, located in the includes folder.